Extracted AbstractSplitter
